+++ test.py
@@ -2,7 +2,7 @@
 
 from math import cos, pi
 from random import randint
-from tkinter import Menu, TclError, messagebox
+from tkinter import TclError, messagebox
 from winsound import Beep
 
 import tkintertools as tkt
@@ -39,7 +39,7 @@ def update(ind: int = 0) -> None:
     """ 进度条更新 """
     bar.load(ind)  # 更新进度条
     if ind < 1:  # 终止条件
-        root.after(1, update, ind+0.0002)  # 迭代执行函数
+        root.after(2, update, ind+0.0002)  # 迭代执行函数
 
 
 def shutdown() -> None:
@@ -50,8 +50,6 @@ def shutdown() -> None:
 
 root = tkt.Tk('tkintertools', 1280, 720, shutdown=shutdown)
 root.minsize(640, 360)
-menu = Menu(root, tearoff=False)
-root.configure(menu=menu)
 canvas_main = tkt.Canvas(root, 1280, 720)
 canvas_main.place(x=0, y=0)
 canvas_doc = tkt.Canvas(root, 1280, 720)
@@ -64,64 +62,67 @@ def sound(): return Beep(600, 100)
 
 
 tkt.Button(
-    canvas_main, 10, 660, 200, 50, 0, '模块文档',
-    command=lambda: (tkt.move(root, canvas_main, 1280*canvas_main.rx, 0, 500, 'rebound'),
-                     tkt.move(root, canvas_doc, 1280*canvas_doc.rx, 0, 500, 'rebound'))
+    canvas_main, 10, 660, 200, 50, text='模块文档',
+    command=lambda: (tkt.move(root, canvas_main, 1280*canvas_main.rx, 0, 500, mode='rebound'),
+                     tkt.move(root, canvas_doc, 1280*canvas_doc.rx, 0, 500, mode='rebound'))
 ).command_ex['press'] = sound
 tkt.Button(
-    canvas_main, 1070, 660, 200, 50, 0, '图像测试',
-    command=lambda: (tkt.move(root, canvas_main, -1280*canvas_main.rx, 0, 500, 'rebound'),
-                     tkt.move(root, canvas_graph, -1280*canvas_graph.rx, 0, 500, 'rebound'))
+    canvas_main, 1070, 660, 200, 50, text='图像测试',
+    command=lambda: (tkt.move(root, canvas_main, -1280*canvas_main.rx, 0, 500, mode='rebound'),
+                     tkt.move(root, canvas_graph, -1280*canvas_graph.rx, 0, 500, mode='rebound'))
 ).command_ex['press'] = sound
 tkt.Button(
-    canvas_doc, 1070, 660, 200, 50, 0, '返回主页',
-    command=lambda: (tkt.move(root, canvas_main, -1280*canvas_main.rx, 0, 500, 'rebound'),
-                     tkt.move(root, canvas_doc, -1280*canvas_doc.rx, 0, 500, 'rebound'))
+    canvas_doc, 1070, 660, 200, 50, text='返回主页',
+    command=lambda: (tkt.move(root, canvas_main, -1280*canvas_main.rx, 0, 500, mode='rebound'),
+                     tkt.move(root, canvas_doc, -1280*canvas_doc.rx, 0, 500, mode='rebound'))
 ).command_ex['press'] = sound
 tkt.Button(
-    canvas_graph, 10, 660, 200, 50, 0, '返回主页',
-    command=lambda: (tkt.move(root, canvas_main, 1280*canvas_main.rx, 0, 500, 'rebound'),
-                     tkt.move(root, canvas_graph, 1280*canvas_graph.rx, 0, 500, 'rebound'))
+    canvas_graph, 10, 660, 200, 50, text='返回主页',
+    command=lambda: (tkt.move(root, canvas_main, 1280*canvas_main.rx, 0, 500, mode='rebound'),
+                     tkt.move(root, canvas_graph, 1280*canvas_graph.rx, 0, 500, mode='rebound'))
 ).command_ex['press'] = sound
 
+t = tkt.Text(
+    canvas_main, 10, 10, 625, 300, radius=20, text=('居中圆角文本框', '竖线光标'), justify='center')
 tkt.Text(
-    canvas_main, 10, 10, 625, 300, 20, ('居中圆角文本框', '竖线光标'), justify='center')
-tkt.Text(
-    canvas_main, 645, 10, 625, 300, 0, ('靠右方角文本框', '下划线光标'), cursor=' _')
+    canvas_main, 645, 10, 625, 300, text=('靠右方角文本框', '下划线光标'), cursor=' _')
 tkt.Entry(
-    canvas_main, 10, 320, 300, 35, 10, ('居中圆角输入框', '点击输入'), justify='center')
+    canvas_main, 10, 320, 300, 35, radius=10, text=('居中圆角输入框', '点击输入'), justify='center')
 tkt.Entry(
-    canvas_main, 970, 320, 300, 35, 0, ('靠右方角输入框', '点击输入'), '•')
+    canvas_main, 970, 320, 300, 35, text=('靠右方角输入框', '点击输入'), show='•')
 
 tkt.Button(
-    canvas_main, 10, 365, 200, 40, 10, '圆角按钮',
+    canvas_main, 10, 365, 200, 40, radius=10, text='圆角按钮',
     command=lambda: tkt.move(
-        canvas_main, label_1, 0, -170 * canvas_main.ry, 500, 'flat')
+        canvas_main, label_1, 0, -170 * canvas_main.ry, 500, mode='flat')
 ).command_ex['press'] = sound
 tkt.Button(
-    canvas_main, 1070, 365, 200, 40, 0, '方角按钮',
+    canvas_main, 1070, 365, 200, 40, text='方角按钮',
     command=lambda: tkt.move(
-        canvas_main, label_2, 0, -170 * canvas_main.ry, 500, 'smooth')
+        canvas_main, label_2, 0, -170 * canvas_main.ry, 500, mode='smooth')
 ).command_ex['press'] = sound
 
 label_1 = tkt.Label(
-    canvas_main, 235, 730, 400, 150, 20, '圆角标签\n移动模式:flat')
+    canvas_main, 235, 730, 400, 150, radius=20, text='圆角标签\n移动模式:flat')
 label_2 = tkt.Label(
-    canvas_main, 645, 730, 400, 150, 0, '方角标签\n移动模式:smooth')
+    canvas_main, 645, 730, 400, 150, text='方角标签\n移动模式:smooth')
 button_1 = tkt.Button(
-    canvas_doc, 1070, 10, 200, 50, 0, '颜色变幻', command=lambda: (button_1.set_live(False), colorful()))
+    canvas_doc, 1070, 10, 200, 50, text='颜色变幻', command=lambda: (button_1.set_live(False), colorful()))
 button_1.command_ex['press'] = sound
 button_2 = tkt.Button(
-    canvas_graph, 10, 10, 200, 50, 0, '绘制图形',  command=lambda: (button_2.set_live(False), draw()))
+    canvas_graph, 10, 10, 200, 50, text='绘制图形',  command=lambda: (button_2.set_live(False), draw()))
 button_2.command_ex['press'] = sound
 
 load = tkt.Button(
-    canvas_main, 540, 365, 200, 40, 0, '加载进度', command=lambda: (update(), load.set_live(False)))
+    canvas_main, 540, 365, 200, 40, text='加载进度', command=lambda: (update(), load.set_live(False)))
 load.command_ex['press'] = sound
 bar = tkt.Progressbar(canvas_main, 320, 320, 640, 35)
 
+tkt.Button(canvas_main, 500, 465, 280, 40, text='字体选择对话框',
+           command=lambda: tkt.askfont(root, lambda font: print(font)))
+
 canvas_doc.create_text(  # 模块说明文档
-    15, 360, text=tkt.__doc__, font=('consolas', 16, 'italic'), anchor='w')
+    15, 360, text=tkt.__doc__, font=('consolas', 15, 'italic'), anchor='w')
 
 try:  # 加载图片
     canvas_graph.create_image(
@@ -129,5 +130,4 @@ try:  # 加载图片
 except TclError:  # 缺少图片
     print('\033[31mLoad tkintertools.png Error\033[0m')
 
-tkt.SetProcessDpiAwareness()  # 设置DPI级别
 root.mainloop()  # 消息事件循环